emr_etl/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
python-dotenv>=1.0
pydantic>=2.7
pandas>=2.2
openpyxl>=3.1
pyxlsb>=1.0
python-dateutil>=2.9
